tier-list-app/pyproject.toml
[project]
name = "tier-list-app"
version = "0.1.0"
description = ""
authors = [
    {name = "Fialco",email = "113672867+Fialco@users.noreply.github.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "invoke (>=2.2.0,<3.0.0)",
    "tkinterdnd2 (>=0.4.3,<0.5.0)",
    "pillow (>=11.2.0,<12.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
coverage = "^7.7.1"
pylint = "^3.3.6"
autopep8 = "^2.3.2"

